Created extra file for UIElements for cleaner file structure
