+++ src/components/PresetUI.tsx
@@ -1,8 +1,8 @@
 import { createSignal, createEffect } from "solid-js";
 import { ToyManager } from "../js/miditoy/ToyManager";
 import { PresetManager } from "../js/PresetManager";
-import { InitToy } from "../js/solidjs/ComponentUtils.jsx";
-import { DetailsFillerCenter, JsonFileUploader, Button } from "../js/solidjs/ComponentUtils.jsx";
+import * as utils from "../js/solidjs/ComponentUtils.js";
+import * as ui from "./UIElements.jsx"
 
 var tManager = new ToyManager();
 var pManager = new PresetManager();
@@ -22,12 +22,12 @@ export default function SetupContainer( props: {channel: number}) {
 
     const ToyChanged = () => {
         console.log("PRESET UI event");
-        toy = InitToy(channel, toy, ToyChanged);
+        toy = utils.InitToy(channel, toy, ToyChanged);
         UpdateUIValues();
     };
 
     function LoadToy() {
-        toy = InitToy(channel, toy, ToyChanged);
+        toy = utils.InitToy(channel, toy, ToyChanged);
     }
 
     function UpdateUIValues() {
@@ -101,7 +101,7 @@ export default function SetupContainer( props: {channel: number}) {
                             onChange={(event) => setPresetName(event.target.value)}
                         />
                         </div>
-                        <Button 
+                        <ui.Button 
                             class="thinButton marginAuto"
                             onClick={() => SaveNewPreset()}
                             label="Save"
@@ -113,7 +113,7 @@ export default function SetupContainer( props: {channel: number}) {
                 <br></br>
                 {RenderAvailablePresets()}
                 <br></br>
-                <JsonFileUploader 
+                <ui.JsonFileUploader 
                 onFileUpload={UploadPreset}
                 />
             </>
@@ -125,18 +125,18 @@ export default function SetupContainer( props: {channel: number}) {
           <div class="">
             {matchingItems().map((item) => (
                 <div class="flexContainer">
-                        <Button
+                        <ui.Button
                             class="width70 thinButton"
                             onClick={() => LoadPreset(item)}
                             label={GetPresetName(item)}
                         />
                     <div class="width30">
-                        <Button 
+                        <ui.Button 
                             class="squareButton"
                             label="DL"
                             onClick={() => DownloadPreset(item)}
                         />
-                        <Button 
+                        <ui.Button 
                             class="squareButton"
                             label="X"
                             onClick={() => DeletePreset(item)}
@@ -151,5 +151,5 @@ export default function SetupContainer( props: {channel: number}) {
     LoadToy();
     UpdateUIValues(); //Get UI Values once at start
     var sumName = "Load/Save Preset";
-    return DetailsFillerCenter("Load/Save Settings", RenderUI());
+    return ui.DetailsFillerCenter("Load/Save Settings", RenderUI());
 }
