+++ src/components/UIElements.jsx
@@ -1,32 +1,4 @@
 import { createSignal, createEffect } from 'solid-js';
-import { ToyManager } from "../miditoy/ToyManager";
-
-const tManager = new ToyManager();
-
-export function InitToy(channel, toy, ToyChanged) {
-    toy = tManager.GetToy(channel);
-    // toy.UnsubscribeFromToyChangedEvent(ToyChanged);
-    if(toy != undefined) {
-        toy.SubscribeToToyChangedEvent(ToyChanged);
-        return toy;
-    }
-}
-
-export function UnsubscribeEvent(targetFunction) {
-    toy.UnsubscribeFromToyChangedEvent(targetFunction);
-}
-
-export function CreateToy(channel, type) {
-    //If toyType changed, create toy, otherwise, just udpate
-    switch(type) {
-        case 0: tManager.CreateEmptyToy(channel);       break;
-        case 1: tManager.CreateGraviBoard(channel);     break;
-        case 2: tManager.CreatePolyDrum(channel);       break;
-        // case 3: tManager.CreateSquareKeyboard(channel, numberOfKeys, startKey); break;
-        default: tManager.CreateEmptyToy(channel);      break;
-    }
-    return tManager.GetToy(channel);
-}
 
 export function DetailsFillerCenter(summeryName, content) {
     return (
@@ -205,19 +177,6 @@ export function JsonFileUploader(props) {
         reader.readAsText(file);
       });
     }
-    
-
-    // if (file) {
-    //   const reader = new FileReader();
-
-    //   reader.onload = (event) => {
-    //     const jsonData = JSON.parse(event.target.result);
-    //     // Pass the uploaded JSON data to the callback function
-    //     props.onFileUpload(file.name, jsonData);
-    //   };
-
-    //   reader.readAsText(file);
-    // }
   };
 
   return (
