+++ src/icons/micions Interface/Licence.txt
@@ -1,0 +1,1 @@
+https://www.svgrepo.com/page/licensing#CC%20Attribution
