+++ src/icons/micions Interface/Source.txt
@@ -1,0 +1,1 @@
+https://www.svgrepo.com/collection/micions-interface-icons/
